feat:(tests-category) add tests layer services category

+++ LevvaCoins.Application/Transactions/Commands/SaveTransactionCommand.cs
@@ -8,11 +8,11 @@ namespace LevvaCoins.Application.Transactions.Commands
     {
         public Guid UserId { get; }
         public string Description { get;}
-        public double Amount { get; }
+        public decimal Amount { get; }
         public TransactionType Type { get; }
         public Guid CategoryId { get; }
 
-        public SaveTransactionCommand(Guid userId, string description, double amount, TransactionType type, Guid categoryId)
+        public SaveTransactionCommand(Guid userId, string description, decimal amount, TransactionType type, Guid categoryId)
         {
             Description = description;
             Amount = amount;
